pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
    "pillow>=11.2.1",
    "pyjwt>=2.10.1",
    "numpy>=2.2.5",
    "opencv-python>=4.11.0.86",
    "weasyprint>=65.1",
    "openpyxl>=3.1.5",
    "python-dotenv>=1.1.0",
    "fpdf>=1.7.2",
    "xlsxwriter>=3.2.3",
]
